requirements.txt
attrs==26.1.0
Automat==25.4.16
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
constantly==23.10.4
cryptography==46.0.7
cssselect==1.4.0
defusedxml==0.7.1
fastjsonschema==2.21.2
filelock==3.25.2
future==1.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.4.0
jmespath==1.1.0
listcrunch==1.0.1
lxml==6.0.3
packaging==26.0
parsel==1.11.0
Protego==0.6.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyDispatcher==2.0.7
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-documentcloud==4.5.0
python-squarelet==0.2.1
PyYAML==6.0.3
queuelib==1.9.0
ratelimit==2.2.1
requests==2.33.1
requests-file==3.0.1
Scrapy==2.15.0
service-identity==24.2.0
six==1.17.0
tldextract==5.3.1
Twisted==25.5.0
typing_extensions==4.15.0
urllib3==2.6.3
w3lib==2.4.1
zope.interface==8.2
